requirements.txt
#cachetools==3.1.1
#google-auth==1.6.3
#six==1.12.0
#slackclient==1.3.1
#urllib3==1.26.5
#websocket-client==0.56.0
PySocks==1.7.1
asn1crypto==1.2.0
beautifulsoup4==4.8.1
cachetools==4.2.2
certifi==2019.9.11
cffi==1.13.0
chardet==3.0.4
cryptography==3.3.2
python-dotenv
flask
geojson==2.5.0
google-api-python-client==1.7.11
google-auth-httplib2==0.0.3
google-auth<2.0dev,>=1.25.0
httplib2==0.19.0
idna==2.8
lxml==4.6.5
nltk==3.6.5
openai
pyOpenSSL==19.0.0
pyasn1-modules==0.2.7
pyasn1==0.4.7
pycparser==2.19
pyowm==2.10.0
requests==2.22.0
rsa==4.7
six>=1.13.0
slack_bolt>=1.6.1
slack_sdk>=3.0
soupsieve==1.9.3
textblob==0.15.3
uritemplate==3.0.0
urllib3
websocket-client
